// src/app/interfaces/IConsentModel.ts
export interface IConsentModel {
  id?: number;
  IamBorrowerCoBorrower?: string;
  agreeEConsent?: boolean;
  firstName?: string;
  lastName?: string;
  email?: string;
  CoborrowerAgreeEConsent?: boolean;
  CoborrowerFirstName?: string;
  CoborrowerLastName?: string;
  CoborrowerEmail?: string;
}