// src/app/interfaces/IBorrowerModel.ts
export interface IBorrowerModel {
  id?: number;
  firstName?: string;
  middleInitial?: string;
  lastName?: string;
  suffix?: string;
  email?: string;
  dateOfBirth?: string;
  socialSecurityNumber?: string;
  maritalStatusId?: number;
  numberOfDependents?: number;
  cellPhone?: string;
  homePhone?: string;
}